refactor: move configuration file to data directory and update references
