+++ server/routes/auth.js
@@ -5,7 +5,7 @@ require("dotenv").config();
 const createOTP = require("../utils/otp");
 const LogConsole = require("../utils/logging");
 const basicIpFilter = require("../middleware/basicIpFilter");
-const config = require("../../config.json");
+const config = require("../../data/config.json");
 
 const router = express.Router();
