+++ server/routes/authConfig.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const config = require("../../config.json");
+const config = require("../../data/config.json");
 
 const router = express.Router();
