+++ server/routes/wake.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const basicIpFilter = require("../middleware/basicIpFilter");
 const sendMagicPacket = require("../logic/sendMagicPacket");
 const LogConsole = require("../utils/logging");
-const config = require("../../config.json");
+const config = require("../../data/config.json");
 
 // Custom middleware to check if the user is authenticated
 const customAuth = (req, res, next) => {
